+++ backend/app/main.py
@@ -1,12 +1,12 @@
 from fastapi import FastAPI, Depends
 from fastapi.exceptions import HTTPException
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 from starlette.middleware.cors import CORSMiddleware
 from starlette.status import HTTP_400_BAD_REQUEST
 
 from . import crud, models
-from .auth import create_access_token, decode_token
+from .auth import create_access_token, get_current_user
 from .database import engine, get_db
 from .routers import users
 
@@ -19,32 +19,9 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-oauth2_schema = OAuth2PasswordBearer(tokenUrl="/token")
 app.include_router(users.router)
 
 models.Base.metadata.create_all(bind=engine)
-
-
-async def get_current_user(
-    token: str = Depends(oauth2_schema),
-    db: Session = Depends(get_db)
-) -> models.User:
-    """Get a user for the token that"s passed in via HTTP headers.
-
-    This method gets the user's token through a `Depends` and processeses it
-    into a `models.User` object (if it's valid).
-
-    Args:
-        token: user's session token ('Authorization' header)
-        db: database connection
-
-    Returns:
-        models.User: user for the token
-
-    Raises:
-        Exception if the token couldn't be decoded or didn't match a user
-    """
-    return decode_token(token, db)
 
 
 @app.post("/token")
